added opponent prediction percentage after it was found

+++ project/monteCarloSim/monteCarloSim.cpp
@@ -99,8 +99,8 @@ Eigen::VectorXf MonteCarloSim::getStateForAgent() const
         applyToArray(opPos, agentStateSize);
     }
     //        applyToArray(goalPos,agentStateSize*2);
-    float xDiff = static_cast<int>(goalPos.x - agentPos.x) / 5.0f;
-    float yDiff = static_cast<int>(goalPos.y - agentPos.y) / 5.0f;
+    float xDiff = (static_cast<int>(goalPos.x) - static_cast<int>(agentPos.x)) / 5.0f;
+    float yDiff = (static_cast<int>(goalPos.y) - static_cast<int>(agentPos.y)) / 5.0f;
     if (xDiff < 0)
     {
         agentGrid[offsetForGoal] = -xDiff;
